pyproject.toml
[tool.poetry]
name = "nsc"
version = "1.0"
description = "Python dependency manager for Nassau Street Capital"
authors = ["Windsor Nguyen"]
license = "MIT"
# homepage = "https://nassaustreet.co"
repository = "https://github.com/windsornguyen/nsc"
keywords = ["finance", "investing", "software", "quant", "princeton"]

[tool.poetry.dependencies]
python = "^3.9"
multidict = "5.2.0a25"
discord-py = "^2.3.2"
openai = "^1.3.5"
setuptools = "^69.0.2"
python-dotenv = "^1.0.0"
tastytrade-sdk = "^1.0.0"
apscheduler = "^3.10.4"
pandas = "^2.1.3"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
poetry-plugin-export = "^1.6.0"
pre-commit = "^3.5.0"


[tool.poetry.dev-dependencies]
# See `ruff help` for more on Ruff's top-level commands.
ruff = "^0.1.5"
pre-commit = "^3.5.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assuming Python 3.9+.
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "B", "Q"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# This is the (only) correct way to style strings (all my homies hate PEP 8)
[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"  # Enforce double quotes for docstrings
inline-quotes = "single"     # Enforce single quotes for inline strings

[tool.ruff.format]
quote-style = "single"       # Default to single quotes for auto-formatting

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
